+++ packages/observable-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Error when unable to track a transaction because we havne't fully recovered from a `stop` event
+
 ### Added
 
 - Emit current finalized and descendants when subscribing to `newBlocks$`